refactor(ui): split _display into focused submodules
Move display helpers into ui/menus/_display/ package with re-exports
from __init__ for backward-compatible imports.
