+++ ui/menus/_display/__init__.py
@@ -1,0 +1,70 @@
+"""Отображение карточек персонажей, рас и характеристик."""
+
+from ui.menus._display._character import (
+    _character_base_race_label,
+    _character_subrace_label,
+    _empty_field_value,
+    _format_proficiency_token_list,
+    _print_character_card,
+    _print_character_proficiencies,
+    _print_characters_list,
+    _print_labeled_field,
+)
+from ui.menus._display._class import (
+    _character_class_label,
+    _character_subclass_label,
+    _format_class_proficiencies,
+    _format_class_skills,
+    _format_feature_uses,
+    _print_class_description,
+    _print_class_features,
+    _print_class_info,
+    _print_class_meta_line,
+    _print_class_summary,
+    _print_features_section_title,
+    _print_subclass_info,
+)
+from ui.menus._display._difficulty import _difficulty_color, _difficulty_label
+from ui.menus._display._race import (
+    _format_bonuses,
+    _print_race_bonuses,
+    _print_race_info,
+)
+from ui.menus._display._stats import (
+    _format_character_stats_compact,
+    _print_final_stat_line,
+    _print_point_buy_cost_table,
+    _print_stats_generation_header,
+)
+
+__all__ = [
+    "_character_base_race_label",
+    "_character_class_label",
+    "_character_subclass_label",
+    "_character_subrace_label",
+    "_difficulty_color",
+    "_difficulty_label",
+    "_empty_field_value",
+    "_format_bonuses",
+    "_format_character_stats_compact",
+    "_format_class_proficiencies",
+    "_format_class_skills",
+    "_format_feature_uses",
+    "_format_proficiency_token_list",
+    "_print_character_card",
+    "_print_character_proficiencies",
+    "_print_characters_list",
+    "_print_class_description",
+    "_print_class_features",
+    "_print_class_info",
+    "_print_class_meta_line",
+    "_print_class_summary",
+    "_print_features_section_title",
+    "_print_final_stat_line",
+    "_print_labeled_field",
+    "_print_point_buy_cost_table",
+    "_print_race_bonuses",
+    "_print_race_info",
+    "_print_stats_generation_header",
+    "_print_subclass_info",
+]
